Fix bug with unreachable PhotoAlbumDetails objects
Each time when user replace the images in a photo album with a new set of images downloaded from Flickr new PhotoAlbumDetails was inserted to main context and previous one become unreachable.
Now, when downloading is finished check for existing albumDetails property on pin instance and update it values, otherwise create a new one PhotoAlbumDetails instance in main context and link it with a pin.